Add Visual Studio Code logo (923 bytes) (#595)
The reference logo is an official SVG, taken from
https://code.visualstudio.com/brand

The tiny SVG is a simplified & minified version of it.
Mostly, there is less shadows / blur / etc. (less SVG <filter>),
so as to make the tiny SVG fit into 1 KiB.

The Android Vector Drawable is a converted version of the tiny SVG, via
Android Studio. Because Vector Drawable doesn’t have equivalents to most
SVG <filter>, it ends up “flatter” than the SVG. However, the logo is
still easily identifiable.